feat : 패배했을 시, 블랙잭 일 시 수익 계산 기능 구현

+++ src/main/java/domain/blackjack/BetAmount.java
@@ -12,7 +12,7 @@ public class BetAmount {
         this.bet = bet;
     }
 
-    public int getPayout(Participant participant, WinStatus winStatus) {
-        return (int) (bet.get(participant) * winStatus.getBetMultiplier());
+    public double getPayout(Participant participant, WinStatus winStatus) {
+        return bet.get(participant) * winStatus.getBetMultiplier();
     }
 }
